Chore: restructuring folders
